learnsets: a word an entry, so a move past 511 can be learnt
A level-up entry packed the move into nine bits and the level into seven. That
holds exactly as long as no move is numbered past 511, and the moves that
landed yesterday go to 923: everything above it was dropped when the learnsets
were written, 1060 entries of them, and no build or test said so. The entry is
a word now, the move in the low half and the level in the high one -- which is
how the reference stores it too -- and the archive was converted in place and
rewritten from the reference. Every reader of it is C, and all four now agree
on one constant.

The same constant sizes the two buffers that were written as numbers: 0x2c in
the move tutor, which held retail's twenty-one moves and not Mr. Rime's thirty,
and 50 in the egg search, which was large enough by luck. That is the fourth
memory bug of this shape in an evening, so it is the fourth one whose size now
comes from the archive rather than from a memory of it.

boot_check learns `sram:`, which puts a save file where the core looks for it.
It does not get past the main menu -- that wants wireless the core does not
emulate -- but it is the first half of starting from a real save.

THE LEDGER

docs/newgold/LEDGER.md is the state of the port, and it is the only place it
is written. tools/newgold/ledger.py reads the counters out of the repository --
the species, move, ability and item ceilings from the headers on both sides,
the trainers from trainers.json, the tests from tests/newgold, the ROM from
the header of the card it just built -- and writes them between the
LEDGER:COUNTS markers here and in README.md. Then it renders docs/index.html
from the Markdown: the heading, the two blocks of bars, and the four tables,
each row's three columns becoming a name, a bar and a chip, with the state
choosing the class and the width. The page had drifted from the ledger within
half an hour of both being written by hand; now it cannot.

Forty-five rows were read against the tree rather than taken on trust, and
twenty of them were saying more than the repository supports. Among them:

  Fairy type has its chart and its 66 species, and no name string and no icon
  the ability field is widened everywhere except the personal record, where
    85 species still truncate
  149 of the 319 abilities do something; the rest are a name
  the Dex text stops at 574, so 467 species have no entry
  Denise is still a vanilla trainer, skipped when Mareanie did not exist
  the Whirl Islands and Cianwood's surf never got konefr's encounters

Each is now its own row with its own state, and test_ledger.py fails if the
counters, the README or the page fall behind. The overall figure moved from
78% to 67% because the ledger stopped rounding up.

Both ROMs build. 198 tests pass.

+++ src/move_relearner.c
@@ -24,7 +24,7 @@ u16 *MoveRelearner_GetEligibleLevelUpMoves(Pokemon *mon, enum HeapID heapID) {
         moves[i] = GetMonData(mon, MON_DATA_MOVE1 + i, NULL);
     }
 
-    u16 *tableFromFile = Heap_Alloc(heapID, LEVEL_UP_LEARNSET_SIZE * sizeof(u16));
+    u32 *tableFromFile = Heap_Alloc(heapID, LEVEL_UP_LEARNSET_SIZE * sizeof(u32));
     u16 *returnTable = Heap_Alloc(heapID, LEVEL_UP_LEARNSET_SIZE * sizeof(u16));
 
     LoadLevelUpLearnset_HandleAlternateForm(species, form, tableFromFile);
